Mask systemd-oomd before the installation
Even though the recent changes tried to keep systemd-oomd under control,
it still insisted on killing the installation script while it was
installing the flatpaked IDEs. This commit will stop and mask the
systemd-oomd service unit to enable the installation script to run on
machines with 4 GiB or less of RAM.

Disabling the systemd-oomd will also avoid surprises when the IDEs try
to analyse or refactor large source codes.

The uninstallation script will try to restore the system's default
configurations for the systemd-oomd service.