Depend on native packages instead of flatpak
Replacing the flatpaked versions of the external IDEs (VSCode, IntelliJ
IDEA, CLion and PyCharm) by the custom native packages required the
following changes:
- updating the .desktop files with those included in the native packages
- removing the hacks used to bypass the Flatpak sandbox

Also, in order to reduce the package's size, the compressed files will
be shipped in their uncompressed form. This was done because the
packaging process compresses the sources and binaries with xz, and
nested xz files compress poorly.

Also, the ICPC user creation script now has an improved debugging
experience.