Fix .desktop trusting process
The .desktop files trusting process, required to execute the files
without manual intervention, was unstable. It was common to see only
half of the files trdusted and the others still requiring manual
intervention.

This commit fixes these instabilities by using systemd-run to login as
the icpc user and launch the required dbus, gdbus and gvfs processes.
The order of the permission and ownership changes has been modified to
avoid problems with the GIO backend not findind the GVFS metadata file.