Prevent VSCode from experimenting on user's machine
The official [FAQ][1] informs that Visual Studio Code will use an
Microsoft-hosted "experimentation framework" to test new features on
user's machines. Also, this system is supposed to be disabled when the
telemetry level is set to "off", yet it can still be forcefully disabled
with the `"workbench.enableExperiments": false` setting on
`.config/Code/User/settings.json`.

[1]: https://code.visualstudio.com/docs/supporting/faq