Fix VSCode insistent keyring creation requests
We now ship a custom `argv.json` that sets the `password store` to the
`basic` mode. This reduces the security of the stored objects, as it
will use a hard-coded key, but this does not concern the project, as we
do not allow Internet connections nor secrets to be saved in the
VSCode's database.